meslimmo/meslimmo/dev-pip-requirements.txt
djangorestframework
markdown
django-rest-auth
djangorestframework-jwt
psycopg2
